@@ -1,55 +1,52 @@
 name: VirtualFlightThrottle
-description: Customizable Virtual Flight-Throttle&#x2F;Controller
+description: VFCS - VFCS Flight Control System
 
-version: 1.0.2+5
+version: 2.0.0+0
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  provider: ^4.0.1
+  async: ^2.8.2
 
-  flutter_speed_dial: ^1.2.5
+  path: ^1.8.0
 
-  card_settings: ^1.16.2
+  path_provider: ^2.0.7
 
-  sqflite: ^1.2.0
+  connectivity: ^3.0.6
 
-  path_provider: ^1.5.1
+  wakelock: ^0.5.6
 
-  connectivity: ^2.0.2
+  url_launcher: ^6.0.15
 
-  wifi_info_flutter: ^1.0.3
+  provider: ^6.0.1
 
-  ping_discover_network: ^0.2.0+1
+  idb_sqflite: ^1.0.1
 
-  flutter_vibrate: ^1.0.0
+  sqlite3_flutter_libs: ^0.5.0
 
-  async: ^2.3.0
+  flutter_speed_dial: ^4.6.6
 
-  flutter_xlider: 3.0.0
+  card_settings: ^3.0.0
 
-  fluttertoast: ^3.1.3
+  wifi_info_flutter: ^2.0.2
 
-  control_pad: ^1.0.1+1
+  flutter_vibrate: ^1.3.0
 
-  dotted_border: ^1.0.4
+  fluttertoast: ^8.0.8
 
-  url_launcher: ^5.4.1
+  dotted_border: ^2.0.0+1
 
-  intl_utils: ^1.0.2
-
-  wakelock: ^0.1.4+1
+  intl_utils: ^2.5.1
 
   flutter_localizations:
     sdk: flutter
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
+  change_app_package_name: ^1.0.0
 
 flutter:
   uses-material-design: true
