@@ -1,8 +1,5 @@
 import 'dart:async';
 
-import 'package:VirtualFlightThrottle/data/data_settings.dart';
-import 'package:VirtualFlightThrottle/data/data_sqlite3_helper.dart';
-
 class NetworkProtocol {
   static const int PASS = 0;
   static const int VALIDATION = -2;
@@ -24,10 +21,10 @@ class NetworkData<T> {
   String toString() => "d${targetInput.toString()}:${value.toString()}";
 
   static NetworkData parsePacket(String packet) {
-    final List<String> splitted = packet.split(":");
-    if (splitted.length == 2) {
-      final int targetInput = int.parse(splitted[0]);
-      if (targetInput == NetworkProtocol.VALIDATION) return NetworkData<String>(targetInput, splitted[1]);
+    final List<String> splatted = packet.split(":");
+    if (splatted.length == 2) {
+      final int targetInput = int.parse(splatted[0]);
+      if (targetInput == NetworkProtocol.VALIDATION) return NetworkData<String>(targetInput, splatted[1]);
       else return NetworkData<int>(targetInput, 0);
     } else return NetworkData<int>(NetworkProtocol.PASS, 0);
   }
@@ -35,12 +32,13 @@ class NetworkData<T> {
 
 class ValidationNetworkData extends NetworkData {
   ValidationNetworkData() : super(NetworkProtocol.VALIDATION,
-      "${AppSettings().settingsMap[SettingsType.USER_NAME].value}/${DateTime.now().millisecondsSinceEpoch}");
+      "VFCS/${DateTime.now().millisecondsSinceEpoch}");
 }
 
 abstract class NetworkAgent {
 
   final String address;
+  final int timeOut;
 
   final Function onSessionKilled;
   bool killed = false;
@@ -48,7 +46,7 @@ abstract class NetworkAgent {
   int _currentValidationReceiveTime = 0;
   int _prvValidationReceiveTime = 0;
 
-  NetworkAgent(this.address, this.onSessionKilled) {
+  NetworkAgent(this.address, this.timeOut, this.onSessionKilled) {
     this._startValidationSession();
   }
 
@@ -73,58 +71,9 @@ abstract class NetworkAgent {
   Future<void> _startValidationSession() async {
     while (!this.killed) {
       this.sendData(ValidationNetworkData());
-      await Future.delayed(Duration(milliseconds: AppSettings().settingsMap[SettingsType.NETWORK_TIMEOUT].value));
+      await Future.delayed(Duration(milliseconds: this.timeOut));
       if (this._prvValidationReceiveTime == this._currentValidationReceiveTime) this.killSession();
     }
   }
-
-}
-
-abstract class NetworkManager {
-
-  bool isConnected = false;
-
-  // ignore: close_sinks
-  StreamController<bool> networkStateStreamController = StreamController<bool>.broadcast();
-
-  NetworkAgent targetNetworkAgent;
-
-  List<NetworkData> _buffer = [];
-
-  Future<bool> checkInterfaceAlive();
-
-  Future<List<String>> findAliveTargetList();
-
-  Future<void> connectToTarget(String deviceAddress, Function() onSessionLost);
-
-  void setConnectedState() {
-    this.isConnected = true;
-    this.networkStateStreamController.add(true);
-    if (AppSettings().settingsMap[SettingsType.AUTO_CONNECTION].value)
-      SQLite3Helper().insertRegisteredDevices(this.targetNetworkAgent.address);
-  }
-
-  void setDisconnectedState() {
-    this.isConnected = false;
-    this.networkStateStreamController.add(false);
-  }
-
-  void disconnectCurrentTarget() {
-    this.setDisconnectedState();
-    this.targetNetworkAgent.killSession();
-    this.isConnected = false;
-  }
-
-  void sendData(NetworkData networkData) {
-    if (!this.isConnected) this._buffer.add(networkData);
-    else if (this._buffer.isNotEmpty) {
-      this._buffer.add(networkData);
-      this._buffer.forEach((val) => this.targetNetworkAgent.sendData(networkData));
-      this._buffer.clear();
-    }
-    else this.targetNetworkAgent.sendData(networkData);
-  }
-
-  NetworkType getInterfaceType();
 
 }
