@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 class Pair<A, B> {
   final A a;
   final B b;
@@ -5,17 +7,33 @@ class Pair<A, B> {
   Pair(this.a, this.b);
 
   @override
-  String toString() => "${a.toString()} : ${b.toString()}";
+  String toString() => "(${a.toString()}, ${b.toString()}";
 }
 
-T getEnumFromString<T>(List<T> enumValues, String sourceString) {
-  T result;
-  enumValues.forEach((val) {if (val.toString() == sourceString) result = val;});
-  return result;
+class Tuple<A, B, C> {
+  final A a;
+  final B b;
+  final C c;
+
+  Tuple(this.a, this.b, this.c);
+
+  @override
+  String toString() => "(${a.toString()}, ${b.toString()}, ${c.toString()})";
 }
+
+R? runGuarded<R>(R Function() body) {
+  try {
+    return body.call();
+  } on Exception {
+    return null;
+  }
+}
+
+T getEnumFromString<T>(List<T> enumValues, String sourceString) =>
+  enumValues.firstWhere((val) => val.toString() == sourceString);
 
 List<T> mixTwoList<T>(List<T> a, List<T> b) {
-  List<T> result = List<T>(a.length+b.length);
+  final List<T> result = List.empty(growable: true);
   for (int idx = 0; idx < a.length; idx++) {
     result[idx*2] = a[idx];
     result[idx*2+1] = b[idx];
@@ -23,6 +41,8 @@ List<T> mixTwoList<T>(List<T> a, List<T> b) {
   return result;
 }
 
-void printWrapped(String text, {int chunk = 800}) {
+String trimString(String text, int length, {int dotLength = 3}) =>
+  text.length > length ? "${text.substring(0, length - 3)}${utf8.decode(List.filled(dotLength, 46))}" : text;
+
+void printWrapped(String text, {int chunk = 800}) =>
   RegExp('.{1, $chunk}').allMatches(text).forEach((match) => print(match.group(0)));
-}
